feat: opcao de selecionar a porta do servidor

+++ app/src-tauri/src/http/server.rs
@@ -13,7 +13,7 @@ use tower_http::services::ServeDir;
 static STOP_TX: Lazy<Mutex<Option<oneshot::Sender<()>>>> = Lazy::new(|| Mutex::new(None));
 
 #[tauri::command(async)]
-pub async fn start_server(app: AppHandle, port: u16) -> () {
+pub async fn start_server(app: AppHandle, port: u64) -> () {
     let (tx, rx) = oneshot::channel::<()>();
     let state = app.state::<Mutex<AppState>>();
     {
